Modify thumb dragging to mimic chrome and FF behavior
Instead of calculating deltaY each frame and adjusting position based on
change, set the indicator to the current mouse position. This prevents
mouse wheel during drag scrolls from being retained (which I kind of
liked) and stops the window from scrolling when your cursor is not
vertically aligned with it (which turned out to feel terrible).