Support elements that don't 100% fill their parents
Scrollbar dimensions and position will now work on elements that don't
have the same width and height as their parent. If they do the option
fillParent=true returns to previous behavior (which does less work on
resizes).

Top and bottom styles on the thumb and padding-top/padding-bottom on the
track will now position the thumb within the track.

Exposed manual update triggers through the refresh() method.