Fix custom dimensions on Firefox
Since Firefox's scrollbar hiding requires the scrollElement to be 100%
width it doesn't work when it doesn't fill the parent. Inserting divs so
that it can have 100% width has proven problematic so the scrollbars
will simply revert back to native. This introduces the useNative option
so that you can maintain scrollStart and scrollStop events with native
scrollbars.