Remove scrollbar's reference to parent styledscroll object to assist GC
Moved requestAnimationFrame update logic into the Scrollbar class since
there was no longer a styledScroll reference.

Improved destroy() to no longer leak memory when the scrollElement is
still attached to the dom.

Moved scrollbar dragging functions from prototype to a new function for
each object so that the callback could be used directly, eliminating the
necessity of a handleEvent method that switched on type.

Micro optimization: changed forEach usage to reverse for loops based on
jsperf results.