Refactor to decrease minified file size, add banner comment
This was primarily an experiment: I was able to reduce minified file
size by 25%.

Refactored repeated for loops into separate functions.
Introduced local variables so they could be mangled (mostly self instead
of this)
Private properties and methods are now preceded by an _ which are then
mangled in the minification process.

Upgraded to gulp 4 for easy sequential tasks.