Support elements that don't fill their parent
Elements can now have a different width and height than their parent be
positioned within them. If a width is set the scrollbar will appear on
the right side of the wrapper so margin-right and right should be used
instead.

Introduced the sameClientWidth, sameClientHeight, and sameDimensions
options to opt IN to the higher performance layout of have the same
dimensions as the parent (sameClientWidth is required for Firefox
support). This replaces the previous customDimensions option.